+++ example/lib/main.dart
@@ -33,8 +33,10 @@ class _MyAppState extends State<MyApp> {
   }
 
   Future<void> _configureDefaultMenus() async {
+    // Deprecated method, start working with the new widgets
+    /*
     await _menuDelegate.configureDefaultMenus({
-      'file': {
+            'file': {
         'additionalItems': [
           {'id': 100, 'label': 'Mi Nuevo Archivo', 'enabled': true},
           {'id': 101, 'label': 'Mi Abrir Especial', 'enabled': true},
@@ -45,8 +47,9 @@ class _MyAppState extends State<MyApp> {
           {'id': 102, 'label': 'Mi Función Personalizada', 'enabled': true},
         ],
       },
-      'hidden': ['format', 'file', 'edit'],
+      'hidden': [],
     });
+    */
   }
 
   @override
@@ -58,10 +61,19 @@ class _MyAppState extends State<MyApp> {
         ),
         child: PlatformMenuBar(
           menus: [
+            IPadEditMenu(
+              onUndo: () => debugPrint('Undo action!'),
+              onRedo: () => debugPrint('Redo action!'),
+            ),
+            IPadFileMenu(),
+            IPadWindowMenu(),
+            IPadViewMenu(),
+            IPadFormatMenu(),
             PlatformMenu(
               label: 'Test Menu',
               menus: [
-                PlatformMenuItem(
+                PlatformMenuItemWithIcon(
+                  icon: CupertinoIcons.plus,
                   label: 'Item 0 (Enabled: $toggledOption)',
                   onSelected: toggledOption
                       ? () => debugPrint("Item 0 selected")
@@ -79,23 +91,45 @@ class _MyAppState extends State<MyApp> {
                     ),
                   ],
                 ),
+
+                // NOTE Item 3 is nested in the same group scope as item 0, so it
+                // will have a leading padding because item 0 has an icon.
+                // To counter this behavior, use PlatformMenuItemGroup for each
+                // section.
                 PlatformMenuItem(
                   label: 'Item 3',
                   onSelected: () => debugPrint("Item 3 selected"),
+                ),
+
+                PlatformMenuItemGroup(
+                  members: [
+                    PlatformMenuItem(
+                      label: 'Item 4',
+                      onSelected: () => debugPrint("Item 4 selected"),
+                    ),
+                    PlatformMenuItem(
+                      label: 'Item 5',
+                      onSelected: () => debugPrint("Item 5 selected"),
+                    ),
+                  ],
                 ),
               ],
             ),
             PlatformMenu(
               label: 'Another Test Menu',
               menus: [
-                PlatformMenuItem(
+                PlatformMenuItemWithIcon(
+                  icon: toggledOption
+                      ? CupertinoIcons.checkmark_alt
+                      : CupertinoIcons.xmark,
                   label: 'Toggled: $toggledOption',
                   onSelected: () => setState(() {
                     toggledOption = !toggledOption;
                   }),
                 ),
                 if (toggledOption)
-                  PlatformMenu(
+                  PlatformMenuWithIcon(
+                    icon: CupertinoIcons.ellipses_bubble,
                     label: "Unlocked Secrets",
                     menus: [
                       PlatformMenuItem(