+++ lib/apple_default_menu_items_enum.dart
@@ -1,2 +1,0 @@
-/// Apple's default menu items on all apps
-enum AppleDefaultMenuItems { appInfo, file, edit, format, view, window, help }