+++ lib/ipados_menu_bar.dart
@@ -3,3 +3,6 @@ This file is used exclusively for exports and to keep a single import
 using the same package name
 */
 export 'ipados_platform_menu_delegate.dart';
+export 'widgets/default_menus/ipad_menus.dart';
+export 'widgets/icon/item_with_icon.dart';
+export 'widgets/icon/submenu_with_icon.dart';