+++ lib/widgets/default_menus/abstract_menu.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/cupertino.dart';
+
+abstract class IPadMenu extends PlatformMenu {
+  IPadMenu({required super.label, required super.menus});
+
+  String get menuId;
+}