+++ lib/widgets/icon/submenu_with_icon.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/cupertino.dart';
+
+class PlatformMenuWithIcon extends PlatformMenu {
+  IconData icon;
+
+  PlatformMenuWithIcon({
+    required this.icon,
+    required super.label,
+    required super.menus,
+  });
+}