+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.0
+
+* Default menus now work like the classic PlatformMenu widgets and are configured at the widget tree.
+  * `IPadFileMenu`
+  * `IPadEditMenu`
+  * `IPadFormatMenu`
+  * `IPadViewMenu`
+  * `IPadWindowMenu`
+* `PlatformMenuItemWithIcon` and `PlatformMenuWithIcon` to pass `IconData` to menu items
+
 ## 0.0.3
 
 * Corrected README.md to correctly show GIF showcase on pub.dev